Do not overwrite sample files

+++ bootstrap.sh
@@ -58,7 +58,10 @@ bundle check &> /dev/null || bundle install --quiet
 
 echo "Copying sample files"
 for sample in $(find config -type f -maxdepth 1 -name '*.sample'); do
-  cp "$sample" "${sample%.*}"
+  if [[ -f "$sample" ]]; then
+    cp "$sample" "${sample%.*}"
+  fi
+  echo "    $sample"
 done
 
 echo "Setup development database"
